A few more tweaks, added Size to the listing, used size_t in a few places, removed a few printfs, etc.
svn path=/trunk/kdebase/kioslave/; revision=23986

+++ pop3/pop3.h
@@ -44,7 +44,7 @@ public:
     * getSize command will emit a signal, which would be harder
     * to trap when doing something like listing a directory.
     */
-  long realGetSize(unsigned int msg_num);
+  size_t realGetSize(unsigned int msg_num);
 
   /**
     *  Send a command to the server, and wait for the one-line-status
